refactor: 요청 데이터 처리 시 DTO를 사용하도록 변경
