+++ src/main/java/roomescape/dto/ReservationRequest.java
@@ -1,28 +1,17 @@
-package roomescape.reservation;
+package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public class Reservation {
-
-    private final Long id;
+public class ReservationRequest {
     private final String name;
     private final LocalDate date;
     private final LocalTime time;
 
-    public Reservation(Long id, String name, LocalDate date, LocalTime time) {
-        this.id = id;
+    public ReservationRequest(String name, LocalDate date, LocalTime time) {
         this.name = name;
         this.date = date;
         this.time = time;
-    }
-
-    public boolean isSameId(Long id) {
-        return this.id.equals(id);
-    }
-
-    public Long getId() {
-        return id;
     }
 
     public String getName() {
